#291: first attempt at externalizing a class and building dist file with esbuild
